requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
requests
plotly
matplotlib
python-dotenv